requirements.txt
fastapi
uvicorn[standard]
pymysql
requests
beautifulsoup4
boto3
python-dotenv
transformers>=4.30.0
torch
